Move testdata to central location
