+++ tools/internal/modinfo/modinfo.go
