+++ tools/testdata/scenarioa/foo/foo.go
