+++ tools/testdata/scenarioa/foo/stage/foo.go
