+++ tools/testdata/scenariob/foo/foo.go
