+++ tools/testdata/scenariob/foo/stage/foo.go
