+++ tools/testdata/scenarioc/foo/foo.go
