+++ tools/testdata/scenarioc/foo/stage/foo.go
