+++ tools/testdata/scenariod/foo/foo.go
