+++ tools/testdata/scenariod/foo/stage/foo.go
