+++ tools/testdata/scenariod/foo/v2/foo.go
