+++ tools/testdata/scenarioe/foo/foo.go
