+++ tools/testdata/scenarioe/foo/go.mod
@@ -1,0 +1,3 @@
+module github.com/Azure/azure-sdk-for-go/tools/testdata/scenarioe/foo
+
+go 1.12
