+++ tools/testdata/scenarioe/foo/stage/foo.go
