+++ tools/testdata/scenarioe/foo/v2/foo.go
