+++ tools/testdata/scenariof/foo/stage/foo.go
