+++ tools/versioner/cmd/root.go
@@ -26,7 +26,7 @@ import (
 	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/tools/apidiff/repo"
-	"github.com/Azure/azure-sdk-for-go/tools/versioner/internal/modinfo"
+	"github.com/Azure/azure-sdk-for-go/tools/internal/modinfo"
 	"github.com/Masterminds/semver"
 	"github.com/spf13/cobra"
 )
