+++ tools/versioner/cmd/root_test.go
@@ -27,7 +27,7 @@ import (
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/tools/apidiff/report"
-	"github.com/Azure/azure-sdk-for-go/tools/versioner/internal/modinfo"
+	"github.com/Azure/azure-sdk-for-go/tools/internal/modinfo"
 )
 
 var newline = "\n"
@@ -228,25 +228,25 @@ func Test_calculateModuleTagMinorV2(t *testing.T) {
 }
 
 func Test_findLatestMajorVersion(t *testing.T) {
-	ver, err := findLatestMajorVersion("../testdata/scenarioa/foo/stage")
+	ver, err := findLatestMajorVersion("../../testdata/scenarioa/foo/stage")
 	if err != nil {
 		t.Fatalf("failed to find LMV: %v", err)
 	}
-	if ver != filepath.Join("..", "testdata", "scenarioa", "foo") {
+	if ver != filepath.Join("..", "..", "testdata", "scenarioa", "foo") {
 		t.Fatalf("bad LMV %s", ver)
 	}
-	ver, err = findLatestMajorVersion("../testdata/scenariod/foo/stage")
+	ver, err = findLatestMajorVersion("../../testdata/scenariod/foo/stage")
 	if err != nil {
 		t.Fatalf("failed to find LMV: %v", err)
 	}
-	if ver != filepath.Join("..", "testdata", "scenariod", "foo", "v2") {
+	if ver != filepath.Join("..", "..", "testdata", "scenariod", "foo", "v2") {
 		t.Fatalf("bad LMV %s", ver)
 	}
-	ver, err = findLatestMajorVersion("../testdata/scenariof/foo/stage")
+	ver, err = findLatestMajorVersion("../../testdata/scenariof/foo/stage")
 	if err != nil {
 		t.Fatalf("failed to find LMV: %v", err)
 	}
-	if ver != filepath.Join("..", "testdata", "scenariof", "foo") {
+	if ver != filepath.Join("..", "..", "testdata", "scenariof", "foo") {
 		t.Fatalf("bad LMV %s", ver)
 	}
 }
@@ -304,12 +304,12 @@ func Test_updateGoModVerB(t *testing.T) {
 }
 
 func cleanTestData() {
-	cmd := exec.Command("git", "clean", "-xdf", "../testdata")
+	cmd := exec.Command("git", "clean", "-xdf", "../../testdata")
 	err := cmd.Run()
 	if err != nil {
 		panic(err)
 	}
-	cmd = exec.Command("git", "checkout", "--", "../testdata")
+	cmd = exec.Command("git", "checkout", "--", "../../testdata")
 	err = cmd.Run()
 	if err != nil {
 		panic(err)
@@ -335,11 +335,11 @@ func Test_theCommandImplMajor(t *testing.T) {
 			return nil, fmt.Errorf("bad prefix '%s'", prefix)
 		}
 		return []string{
-			"tools/versioner/testdata/scenariob/foo/v1.0.0",
-			"tools/versioner/testdata/scenariob/foo/v1.1.0",
+			"tools/testdata/scenariob/foo/v1.0.0",
+			"tools/testdata/scenariob/foo/v1.1.0",
 		}, nil
 	}
-	pkg, err := filepath.Abs("../testdata/scenariob/foo/stage")
+	pkg, err := filepath.Abs("../../testdata/scenariob/foo/stage")
 	if err != nil {
 		t.Fatalf("failed to get absolute path: %v", err)
 	}
@@ -347,19 +347,19 @@ func Test_theCommandImplMajor(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed: %v", err)
 	}
-	const expected = "tools/versioner/testdata/scenariob/foo/v2.0.0"
+	const expected = "tools/testdata/scenariob/foo/v2.0.0"
 	if tag != expected {
 		t.Fatalf("bad tag, expected '%s' got '%s'", expected, tag)
 	}
-	b, err := ioutil.ReadFile("../testdata/scenariob/foo/v2/go.mod")
+	b, err := ioutil.ReadFile("../../testdata/scenariob/foo/v2/go.mod")
 	if err != nil {
 		t.Fatalf("failed to read go.mod file: %v", err)
 	}
-	const after = "module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenariob/foo/v2\n\ngo 1.12\n"
+	const after = "module github.com/Azure/azure-sdk-for-go/tools/testdata/scenariob/foo/v2\n\ngo 1.12\n"
 	if !strings.EqualFold(string(b), after) {
 		t.Fatalf("bad go.mod file, expected '%s' got '%s'", after, string(b))
 	}
-	if !fileExists(filepath.Join("../testdata/scenariob/foo/v2", "CHANGELOG.md")) {
+	if !fileExists(filepath.Join("../../testdata/scenariob/foo/v2", "CHANGELOG.md")) {
 		t.Fatal("expected changelog in scenariob")
 	}
 }
@@ -373,10 +373,10 @@ func Test_theCommandImplMinor(t *testing.T) {
 			return nil, fmt.Errorf("bad prefix '%s'", prefix)
 		}
 		return []string{
-			"tools/versioner/testdata/scenarioa/foo/v1.0.0",
+			"tools/testdata/scenarioa/foo/v1.0.0",
 		}, nil
 	}
-	pkg, err := filepath.Abs("../testdata/scenarioa/foo/stage")
+	pkg, err := filepath.Abs("../../testdata/scenarioa/foo/stage")
 	if err != nil {
 		t.Fatalf("failed to get absolute path: %v", err)
 	}
@@ -384,19 +384,19 @@ func Test_theCommandImplMinor(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed: %v", err)
 	}
-	const expected = "tools/versioner/testdata/scenarioa/foo/v1.1.0"
+	const expected = "tools/testdata/scenarioa/foo/v1.1.0"
 	if tag != expected {
 		t.Fatalf("bad tag, expected '%s' got '%s'", expected, tag)
 	}
-	b, err := ioutil.ReadFile("../testdata/scenarioa/foo/go.mod")
+	b, err := ioutil.ReadFile("../../testdata/scenarioa/foo/go.mod")
 	if err != nil {
 		t.Fatalf("failed to read go.mod file: %v", err)
 	}
-	after := "module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenarioa/foo" + newline + newline + "go 1.12" + newline
+	after := "module github.com/Azure/azure-sdk-for-go/tools/testdata/scenarioa/foo" + newline + newline + "go 1.12" + newline
 	if !strings.EqualFold(string(b), after) {
 		t.Fatalf("bad go.mod file, expected '%s' got '%s'", after, string(b))
 	}
-	if !fileExists(filepath.Join("../testdata/scenarioa/foo", "CHANGELOG.md")) {
+	if !fileExists(filepath.Join("../../testdata/scenarioa/foo", "CHANGELOG.md")) {
 		t.Fatal("expected changelog in scenarioa")
 	}
 }
@@ -410,10 +410,10 @@ func Test_theCommandImplPatch(t *testing.T) {
 			return nil, fmt.Errorf("bad prefix '%s'", prefix)
 		}
 		return []string{
-			"tools/versioner/testdata/scenarioc/foo/v1.0.0",
+			"tools/testdata/scenarioc/foo/v1.0.0",
 		}, nil
 	}
-	pkg, err := filepath.Abs("../testdata/scenarioc/foo/stage")
+	pkg, err := filepath.Abs("../../testdata/scenarioc/foo/stage")
 	if err != nil {
 		t.Fatalf("failed to get absolute path: %v", err)
 	}
@@ -421,19 +421,19 @@ func Test_theCommandImplPatch(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed: %v", err)
 	}
-	const expected = "tools/versioner/testdata/scenarioc/foo/v1.0.1"
+	const expected = "tools/testdata/scenarioc/foo/v1.0.1"
 	if tag != expected {
 		t.Fatalf("bad tag, expected '%s' got '%s'", expected, tag)
 	}
-	b, err := ioutil.ReadFile("../testdata/scenarioc/foo/go.mod")
+	b, err := ioutil.ReadFile("../../testdata/scenarioc/foo/go.mod")
 	if err != nil {
 		t.Fatalf("failed to read go.mod file: %v", err)
 	}
-	after := "module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenarioc/foo" + newline + newline + "go 1.12" + newline
+	after := "module github.com/Azure/azure-sdk-for-go/tools/testdata/scenarioc/foo" + newline + newline + "go 1.12" + newline
 	if !strings.EqualFold(string(b), after) {
 		t.Fatalf("bad go.mod file, expected '%s' got '%s'", after, string(b))
 	}
-	if !fileExists(filepath.Join("../testdata/scenarioc/foo", "CHANGELOG.md")) {
+	if !fileExists(filepath.Join("../../testdata/scenarioc/foo", "CHANGELOG.md")) {
 		t.Fatal("expected changelog in scenarioc")
 	}
 }
@@ -448,7 +448,7 @@ func Test_theCommandImplNewMod(t *testing.T) {
 		}
 		return []string{}, nil
 	}
-	pkg, err := filepath.Abs("../testdata/scenariof/foo/stage")
+	pkg, err := filepath.Abs("../../testdata/scenariof/foo/stage")
 	if err != nil {
 		t.Fatalf("failed to get absolute path: %v", err)
 	}
@@ -456,19 +456,19 @@ func Test_theCommandImplNewMod(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed: %v", err)
 	}
-	const expected = "tools/versioner/testdata/scenariof/foo/v1.0.0"
+	const expected = "tools/testdata/scenariof/foo/v1.0.0"
 	if tag != expected {
 		t.Fatalf("bad tag, expected '%s' got '%s'", expected, tag)
 	}
-	b, err := ioutil.ReadFile("../testdata/scenariof/foo/go.mod")
+	b, err := ioutil.ReadFile("../../testdata/scenariof/foo/go.mod")
 	if err != nil {
 		t.Fatalf("failed to read go.mod file: %v", err)
 	}
-	after := "module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenariof/foo" + newline + newline + "go 1.12" + newline
+	after := "module github.com/Azure/azure-sdk-for-go/tools/testdata/scenariof/foo" + newline + newline + "go 1.12" + newline
 	if !strings.EqualFold(string(b), after) {
 		t.Fatalf("bad go.mod file, expected '%s' got '%s'", after, string(b))
 	}
-	if fileExists(filepath.Join("../testdata/scenariof/foo", "CHANGELOG.md")) {
+	if fileExists(filepath.Join("../../testdata/scenariof/foo", "CHANGELOG.md")) {
 		t.Fatal("unexpected changelog in scenariof")
 	}
 }
