--- tools/versioner/testdata/scenarioa/foo/stage/go.mod
@@ -1,3 +1,0 @@
-module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenarioa/foo
-
-go 1.12
