--- tools/versioner/testdata/scenarioc/foo/stage/go.mod
@@ -1,3 +1,0 @@
-module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenarioc/foo
-
-go 1.12
