--- tools/versioner/testdata/scenariod/foo/go.mod
@@ -1,3 +1,0 @@
-module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenariod/foo
-
-go 1.12
