--- tools/versioner/testdata/scenarioe/foo/stage/go.mod
@@ -1,3 +1,0 @@
-module github.com/Azure/azure-sdk-for-go/tools/versioner/testdata/scenarioe/foo
-
-go 1.12
